Switch market selection to orders
